change vsupply to be a uint16 to enable reporting of voltages higher than 25.5V fixes #294 closes #295

+++ sw/airborne/subsystems/electrical.h
@@ -5,7 +5,7 @@
 
 struct Electrical {
 
-  uint8_t vsupply; /* supply in decivolts */
+  uint16_t vsupply; /* supply in decivolts */
   int32_t current; /* current in milliamps */
   int32_t consumed; /* consumption in mAh */
